requirements.txt
# local package
-e .

# external requirements
click~=8.0.1
python-dotenv>=0.5.1
pandas
click
pandas
plotly
src~=0.1.0
numpy
setuptools~=57.0.0
lxml
python-rapidjson
dnsdb2~=1.1.3
pytz~=2021.1
python-dateutil~=2.8.1
